Big Changes
I realized i had made some mistakes and had to correct. modulus needed to be assigned to 1 so that it could be multiplied by 10. the iteration needed to start at i=0 while it was less than or equal to the number of digits. and i also changes the last if statement to say it was not equal to 0 rather than greater than 0